requirements.txt
chainlit==0.7.700
gTTS==2.4.0
langchain==0.0.345
matplotlib==3.5.1
networkx==3.1
pygame==2.5.2
python-dotenv==1.0.0
Requests==2.31.0
